backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
psycopg2-binary>=2.9,<3.0
celery>=5.4,<6.0
redis>=5.0,<6.0
segno>=1.6,<2.0
reportlab>=4.2,<5.0
XlsxWriter>=3.2,<4.0
google-i18n-address>=3.1,<4.0
phonenumbers>=8.13,<9.0
python-dotenv>=1.0,<2.0
gunicorn>=22.0,<23.0
ruff>=0.5,<1.0
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
